backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
PyJWT==2.8.0
bcrypt==4.1.3
APScheduler==3.10.4
scapy==2.5.0
wakeonlan==3.1.0
paramiko==3.4.0
httpx==0.27.0
python-multipart==0.0.9
cryptography==42.0.5
netifaces==0.11.0
